feat(H3): embed Navigator SPA in kernel — served from bare metal on port 3000
tools/bundle_webapp.py
  Walks slsos-sim/dist/ and generates kernel/webapp_bundle.c as a C array
  table.  Run 'make bundle' to regenerate after UI changes.

kernel/bundle.h + kernel/webapp_bundle.c
  Compiled-in static asset table (687 KiB: index.html + JS + CSS).
  bundle_find(path) does O(n) path lookup, returns BundleAsset*.
  '/' and '/index.html' both resolve to the SPA root.

net/tcp.{h,c}  tcp_send len uint16_t → uint32_t
  Required to transmit the 586 KB JS bundle (exceeds uint16_t max).
  Existing callers unaffected — uint16_t values implicitly widen.

net/http.c  http_respond_raw + bundle routing
  New http_respond_raw(conn, mime, data, uint32_t len) for binary/large assets.
  http_route checks bundle_find() before webapp_find() for every GET.

Makefile  'make bundle' target
  Builds slsos-sim, runs bundle_webapp.py, writes kernel/webapp_bundle.c.

Verified:
  GET /             → HTTP 200  419 bytes  (index.html)
  GET /assets/*.js  → HTTP 200  599687 bytes
  GET /assets/*.css → HTTP 200  62694 bytes
  GET /api/health   → HTTP 200  JSON (API still works)

+++ net/tcp.h
@@ -50,7 +50,7 @@ void tcp_handle_segment(struct IPv4Header* ip, struct TCPHeader* seg, uint16_t s
 // Called by socket layer
 int  tcp_listen(uint16_t port);             // returns conn_id or -1
 int  tcp_accept(int listen_id);             // blocks until connection; returns new conn_id
-int  tcp_send(int conn_id, const void* buf, uint16_t len);
+int  tcp_send(int conn_id, const void* buf, uint32_t len);
 int  tcp_recv(int conn_id, void* buf, uint16_t max_len);  // returns bytes read
 void tcp_close(int conn_id);
